rename Modules to CourseOverview, move CourseOverview and Lesson into organisms, add utils folder
