+++ src/client/src/components/organisms/course/index.ts
@@ -1,1 +1,3 @@
 export * from "./CourseList";
+export * from "./CourseOverview";
+export * from "./Lesson";
