+++ src/client/src/util/course/getCumulativeLessonIndex.ts
