+++ src/client/src/util/course/index.ts
@@ -1,0 +1,1 @@
+export * from "./getCumulativeLessonIndex";
